[IMP] auth_oauth: move `fragment_to_query_string` in tools
Purpose
=======
`fragment_to_query_string` is used in the social module but now show
a warning because it does not depend on `auth_oauth`, so we move the
function to the "tools" folder.

Task-5491124
